delegate shared model method to model concern
